adds budget model and migration

+++ budgetbackend/db/migrate/20200414235325_create_budgets.rb
@@ -2,7 +2,7 @@ class CreateBudgets < ActiveRecord::Migration[6.0]
   def change
     create_table :budgets do |t|
       t.string :title
-      t.int :total
+      t.decimal :total
       t.belongs_to :user, null: false, foreign_key: true
 
       t.timestamps
